#138 separate detail list item mapper
